requirements.txt
google-genai>=1.73.1
flask>=3.0.0
flask-cors>=4.0.0
python-docx>=1.1.0
pdfplumber>=0.11.0
PyMuPDF>=1.24.0
openpyxl>=3.1.0
pandas>=2.2.0
requests>=2.31.0
selenium>=4.18.0
webdriver-manager>=4.0.0
schedule>=1.2.0
pywin32>=306
beautifulsoup4>=4.12.0
python-dateutil>=2.9.0
tabulate>=0.9.0
tqdm>=4.66.0
chromadb>=0.5.0
pywebview>=5.1.0
waitress>=3.0.0
DrissionPage>=4.0.0
pytesseract>=0.3.10
Pillow>=10.2.0
indic-transliteration>=2.3.5
pydantic>=2.6.0
groq>=0.5.0
